chore: Update deploy scripts in package.json
Updated the existing deploy-dev script to be deploy-beta and point to
Vercel's preview environment. Likewise, the deploy-prod script has been
added to point to Vercel's production environment.

This changes streamlines the process in which developers can quickly
and easily build and deploy to the target environment.

The scripts have been fully tested and work as intended.